Fix size computation for semigroups over GF(p^q)
This fixes issue #91, but only by enforcing the (presumably) correct
representation for the matrices generating the Schutzenberger groups.

I believe this should be taken care of by migrating groups code to
use MatrixObjs

This also introduces LambdaIdentity and LambdaBound attributes
for acting semigroups.
* LambdaIdentity returns a function that
  given the rank of the H class as input,
  returns the identity element of the
  Schutzenberger group.
* LambdaBound returns an enumeration bound
  given the rank of the H class.

These were introduced to slightly clean up LambdaOrbSchutzGroup,
but not yet implemented for all types of semigroups.