Make IsomorphismMatrixSemigroup return an isomorphism
In the case where a change of ground fields is requested, for now
compose an isomorphism into a transformation semigroup with an
isomorphism into a matrix semigroup over the requested ground field.

This very probably leaves a lot to be improved upon if this functionality
is ever supposed to be used in any seriousness since it blows up the
dimensions of the matrices.