Fix and clean up various functions in semirms.gi
For most Rees 0-matrix semigroups and Rees matrix semigroups,
the rows are [1..m] and the columns are [1..n] for some numbers
m and n. However, a subsemigroup of such a semigroup may be
a Rees (0-)matrix semigroup in its own right, defined with the same
matrix but whose rows and columns are subsets of [1..m] and
[1..n].

- The methods for RZMSDigraph, RZMSConnectedComponents,
RZMSNormalization, RMSNormalization, and IsInverseSemigroup
previously assumed that the rows and columns of any Rees (0-)
matrix semigroup were equal to [1..m] and [1..n].

- RZMSNormalization initially only worked for a Rees 0-matrix
semigroup over a group in the category IsGroup. Now any
underlying semigroup is allowed.

- Code coverage is improved, mostly standard tests are reclaimed
from the extreme tests.

This commit also cleans up the files semirms.gi and semirms.tst.