Fix MatrixEntries (Issue #164)
MatrixEntries fails for some Rees 0-matrix semigroups whose
matrices contain 0, since MatrixEntries removes duplicates and
sorts using Union, however 0 is not comparable with all
semigroup elements.

This issue is resolved by using Unique to remove duplicates,
putting 0 (if present) at the start of the list, and then sorting
the remaining elements (which are all from the same
semigroup, and are hence sortable).

Resolves: #164