Declare and document MinimalXGeneratingSet attrs
Although I have no intention currently of installing methods for
MinimalSemigroupGeneratingSet, MinimalMonoidGeneratingSet, etc,
I think it is useful to have these attributes declared. In the case that
something is known to satisfy IsMonogenicX, we can set its
MinimalXGeneratingSet to be that generator.

This can be useful - for instance, once set, it can be used in the
Size method for a monogenic transformation semigroup.

This commit also introduces documentation for these attributes.