Update Size for a monogenic transformation semigroup
This commit removes some unreachable code - the index of a
transformation is always strictly positive.

This commit also removes the use of IrredundantGeneratingSubset
and replaces it with MinimalSemigroupGeneratingSet. Since there are
no methods installed for this, it is now required as a filtered. However
this should not be a problem since MinimalSemigroupGeneratingSet
is set whenever a semigroup can learn that it is monogenic.

The filter of this method is bumped by 4, so that the filter is used over
those methods for acting (regular) semigroups.