Introduce Size for a monogenic transformation monoid
This method is almost the same as that for the size of a monogenic
transformation monoid, however there is a slight complication in that
sometimes the monoid generator generates the One of the collection,
whereas sometimes it doesn't. Therefore we need to be careful with
this case to avoid an off-by-one error.