Update IsMonogenicSemigroup immediate method
The MinimalSemigroupGeneratingSet is now set if this immediate
method returns true.

Also, this immediate method is extended, so even if the elements of the
semigroup can not be easily compared, if it happens that there is only
one generator, the true result is set nonetheless (since no comparison
is needed).

This change requires a test in semicons.tst to be updated.