Update IsMonogenic[Inverse]Semigroup
Whenever IsMonogenic[Inverse]Semigroup returns true, it now also
sets the Minimal[Inverse]SemigroupGeneratingSet of the [inverse]
semigroup.

The D-classes of a monogenic [inverse] semigroup are all trivial,
except possibly the minimal ideal which is a cyclic group. If the
D-classes are known, and their sizes are not as they should be,
then the function now returns false before continuing to the final
stage of the code.