Fix issue #192 (RepresentativeOfMinimalIdealNC)
If the size of a semigroup is known and it is less than the number of vertices
required in the action graph for RepresentativeOfMinimalIdealNC, then
try next method.

Similarly, if the number of vertices in the action graph is unfeasibly large,
then there is no point proceeding with the action graph, and so we should
try next method.

This change introduces these changes, and the required code coverage.