Resolve Issue 219
There was an old and incorrect method for InjectionPrincipalFactor
that was earlier removed in stable-2.8 but not in unstable-3.0
(for some reason). This method is removed, tests added, and an
additional check is added to the proper method for
InjectionPrincipalFactor which checks if an element is in the
D-class in the injection and returns fail if it is not.