Reintroduce Position(semigroup, element)
With the meaning of PositionCurrent, also introduce the C function
semi_obj_get_en_semi_no_init which returns 0L if the en_semi
of the semigroup Obj as not been initialised. This is then used
in every C function asking for current information, so that the
data structures are not initialised.

Also introduced IsFullyEnumerated operation to check if
an enumerable semigroup has been fully enumerated, and
various other cleaning up etc.