General clean up of the kernel module (#229)
This PR contains an extensive rewrite of the kernel module so that it is simpler, better organised, more efficient, and more robust. Corresponding changes were also made in the GAP code as several functions were renamed (all those starting SEMIGROUP_). The other main change is that the GenericSemigroupData object has been removed, it was a bad idea. Also the representation IsEnumerableSemigroupRep is introduced and corresponding methods for it are also introduced. This is so that we can clearly differentiate those methods to which the Froidure-Pin algorithm ought to apply (semigroups of boolean matrices, for example), and those where it does not (fp semigroups).