Fix: clarify different types of Green’s classes and semigroups
Namely, remove IsSemigroupWithInverseOp since it is equivalent to
IsInverseSemigroup and IsGeneratorsOfInverseSemigroup, and introduce
IsInverseActingSemigroupRep and IsRegularActingSemigroupRep (which are
categories). This is to fix Issue #237, since some methods for
regular/inverse acting semigroups and their Green's classes were
incorrect if they learned after they were created that they are
regular/inverse. For example, previously if we calculated the D-classes
of a partial perm semigroup, then ask IsInverseSemigroup, then the
ordering of the D-classes differed from the ordering used by
PartialOrderOfDClasses (for instance) since there is a specialised
method for PartialOrderOfDClasses for an inverse semigroup. This commit
resolves this issue by not allowing a semigroup or monoid to use the
specialist methods where this is an issue, unless they are created
knowing they are regular or inverse.