Fix: MinimalIdealGeneratingSet method
This is the method for ideals with an ideal generating set, previously
it called MaximalDClasses(I), and returned the list of representatives.
This is only correct if I is regular, otherwise it will return a too large
generating set. Also MaximalDClasses(I) currently involves first
calculating a generating, and if the generators of I as an ideal
is very large, then this cased problems. So I replaced it with a
method which finds the maximal D-classes of the supersemigroup
of I contained in I, and returns a list of their representatives.