Add free congruences code, and doc
This is to allow us to define congruences on free semigroups, this
shouldn't be necessary since a congruence on a free semigroup is really
the same as an fp semigroup but there is no congruence in the library
definition of an fp semigroup. This is temporary until a better solution
is found.