semigrp: Introduce Closure(Inverse)Monoid
This commit introduces ClosureMonoid, and ClosureInverseMonoid.
Previously we only had ClosureSemigroup and ClosureInverseSemigroup.
This was necessary because sometimes when using ClosureSemigroup, an
extra element would be in the returned result (an identity), due to
incrementally added elements sometimes resulting in a monoid being
returned inappropriately (for example, when adding identity partial
perms, if an intermediate semigroup became a monoid, then the final
result was a monoid, when it shouldn't have been). This commit resolves
this issue.

ClosureSemigroupNC and ClosureInverseSemigroupNC are renamed
ClosureSemigroupOrMonoidNC and ClosureInverseSemigroupOrMonoidNC.

SEMIGROUPS.AddGenerators is replaced by
SEMIGROUPS.ClosureSemigroupDestructive, due to changes in the C++ code.
Code using AddGenerators and incrementally adding new generators is
mostly replaced with a call to ClosureSemigroup (etc) since this is now
done in the C++ closure method.