kernel: remove incorrect INTOBJ_INT
This commit resolves Issue #284, as CAJ correctly points out, the
variable pos is already a GAP integer object, and so the INTOBJ_INT in
line 1091 was incorrect. I guess this didn't cause any problems because
the result of INTOBJ_INT(pos) is usually a large number. The second
argument of the function fropin, which this issue affects, is the limit,
so using a large argument will probably just have enumerated the
semigroup to the end.