Fixed dead links
It seems that all the URLs were pointed to gap-systems.github.io instead of the correct gap-packages.github.io. There is a legitimate argument to be made that they should point to the gap-systems.org pages, but I felt like the gap-packages.github.io would be better since you are looking to install the package and not necessarily read about it.