semigrp: Set GensOfInvMonoid before GensOfInvSgp
This removes some diffs in the GAP manual examples that occur when the
Semigroups package is loaded. As a result of this change, I had to
change some of the Semigroups package tests, but not many.

Additionally, for GeneratorsOfInverseSemigroup, we always add the One if
the semigroup does not have easily comparable elements (we never added
it before - which was a bug). Also, we add it to the end of the list of
GeneratorsOfInverseMonoid, rather than to the beginning.