build: make libsemigroups a subpackage
This commit makes libsemigroups a proper subpackage (in the sense of
autoconf/make) of Semigroups, and makes configure detect the version of
libsemigroups required and present, to give more meaningful error
messages. With libsemigroups being a subpackage there is no longer any
need to run configure in libsemigroups to generate the file
libsemigroups-config.h since this is done automatically because
libsemigroups is a subpackage. Also we do not have to keep track of the
sources of libsemigroups within the Semigroups package build system.