semieunit: improve McAlisterTripleSemigroup, add synonyms
Replaced a check involving IsSubgroup and AutomorphismGroup by one which
checks if generators are automorphism. Checking generators is easier
than checking for a subgroup. Calculating the automorphism group of a
digraph is unnecessary.

Added some attribute synonyms since for attributes of the form
McAlisterTripleX because the names are long. Synonyms are of the form
MTSX.