semieunit: make these acting semigroups, improve IsomorphismSemigroup
This commit fits McAlisterTripleSemigroups into the acting semigroups
framework (previously they were classed as enumerable).

Furthermore it implements an improvement to IsomorphismSemigroup when
used to find an isomorphism from an E-unitary inverse semigroup to a
McAlisterTripleSemigroup.

Finally, there is now a McAlisterTripleSubsemigroup attribute and
similar functionallity for these objects as for
McAlisterTripleSemigroups.