attr: Add NambooripadPartialOrder & Leq methods
The NambooripadPartialOrder is a generalization of the natural partial
order for inverse semigroups to the class of regular semigroups. These
functions are modelled on the fallback methods for NaturalPartialOrder
and NaturalLeqInverseSemigroup. As I do not have a current use for this,
I haven't specialized the methods for any particular representations of
regular semigroups.