Don't pass immutable plists to SET_TYPE_OBJ
First off, whether the plist is immutable or not does not matter anyway, as
the tnum is discarded. Secondly, on the GAP level, Objectify rejects
immutable lists; while the check for that is currently done on the GAP
language level, it might in the future be done on the kernel level, too; so
better to play it safe.

Also, don't bother to retype to fancy plist TNUMs like  T_PLIST_TAB_RECT, as
they will be discarded.