prop: rename 'Decomposable' semigps to 'Surjective'
In my viva, Nik Ruskuc pointed out that decomposable is
not a good name for a semigroup S that satisfies S^2 = S.
Instead, he suggested `surjective' as a better name (along
with several others). It means sense because a semigroup
satisfies S^2 = S if its multiplication is surjective. This
is the name that I have moved to in my PhD thesis.

Since `IsDecomposableSemigroup` was never included in a
released version of the package, there should be no problems
with renaming it.