Make McAlister triple semigroups acting (#507)
* semieunit: remove error message from OneImmutable

Changed method OneImmutable (for McAlisterTripleSemigroups) so that
returns fail instead of an error when the semigroup is not a monoid. As
an example, the error message was causing an issue when
IsomorphismSemigroup(IsTransformationSemigroup, S) was called on a
McAlisterTripleSemigroup.

* semieunit: utilize DirectProduct in EUnitaryInverseCover

EUnitaryCover was written before DirectProduct was properly implemented
for the relevant types of semigroups, and used its own method to
construct direct products. This method is less effective and now
obsolete so it has been replaced.

* semieunit: improve McAlisterTripleSemigroup, add synonyms

Replaced a check involving IsSubgroup and AutomorphismGroup by one which
checks if generators are automorphism. Checking generators is easier
than checking for a subgroup. Calculating the automorphism group of a
digraph is unnecessary.

Added some attribute synonyms since for attributes of the form
McAlisterTripleX because the names are long. Synonyms are of the form
MTSX.

* semieunit: make these acting semigroups, improve IsomorphismSemigroup

This commit fits McAlisterTripleSemigroups into the acting semigroups
framework (previously they were classed as enumerable).

Furthermore it implements an improvement to IsomorphismSemigroup when
used to find an isomorphism from an E-unitary inverse semigroup to a
McAlisterTripleSemigroup.

Finally, there is now a McAlisterTripleSubsemigroup attribute and
similar functionallity for these objects as for
McAlisterTripleSemigroups.

* semieunit: fix tests

* semieunit: add methods for  LeftOne, RightOne

* semieunit: fix tests and improve code coverage

Fix two tests which have different (but mathematically correct) output
when acting methods are enabled/disabled. Improve code coverage in main/setup.gi.
Noticed that IdempotentCreator method may need to be fixed.

* fix