isorms: Rephrase commit
CanonicalMatrix -> ReesZeroMatrixSemigroupCanonicalLabelling
Add OnReesZeroMatrixSemigroups

I'm not sure if OnReesZeroMatrixSemigroups is satisfactory. It's
possible for the user to apply it using a permutation which doesn't
actually act on the space of RZMS over whichever group - it could
then cause an error by not returning a valid matrix to construct an
RZMS. The other issue it would be very difficult for a user to define
permutation which acts on this space and especially hard to define one
which corresponds to some isomorphism of a triple they have in mind.
The easiest way I can think to test whether the permutation corresponds
to some isomorphism by a triple is to construct the group of all these
permutations and check membership (sounds bad).