conditional installation of libsemigroups
nothing changes if there is no external libsemigroups
known to pkg-config and matching the version test;
there is an extra option --with-included-libsemigroups to force
the use of the included library (it is needed to "make dist"
including libsemigroups). The included libsemigroups
will also be used if no external one is good enough.

Dealing with libsemigroups is moved into m4/ subdir macro.