Avoid using attributes Matrix, Rows, Columns
and also their testers and setters. Instead use one of these:
- RowsOfReesZeroMatrixSemigroup
- RowsOfReesMatrixSemigroup
- ColumnsOfReesZeroMatrixSemigroup
- ColumnsOfReesMatrixSemigroup
- MatrixOfReesZeroMatrixSemigroup
- MatrixOfReesMatrixSemigroup

Right now, in GAP 4.11 and master branch, there are just synonyms for
the attributes they replace, but this may change in the future (namely
in order to ensure that Matrix on a single argument is a plain operation,
not an attribute, as for use as a matrix *constructor* function, it is not
desirable to have it an attribute).