Make sure to include GAP headers before libsemigroups
This way, #defines in libsemigroups don't affect GAP headers. For example, GAP
may typedef BOOL in the future, and right now that would break Semigroups
because libsemigroups #defines BOOL to something. With this patch, that
problem goes away.

It also seems in general useful to have a fixed order; this way, *if* things
break in either GAP or libsemigroups due to a change in the respective other,
at least they will break consistently.