Only declare DirectProductOp/SmallGeneratingSet when required
In GAP versions before 4.12, these are only declared for IsGroup.
This means that we have to separately declare them for IsSemigroup.
This then gives warnings when further methods are installed for
IsGroup, because they don't know which declaration they belong to.
Not a huge problem, but one with an easy fix.

The GAP master branch does, and GAP 4.12 will,
declare these things for IsSemigroup instead. Therefore we should
only make these declarations when GAP has not already made them.
Eventually, the package will presumably require a new enough GAP
version that all of this code can be deleted.