Copy the matrix arg in NewMatrixOverFiniteField
`ImmutableMatrix` is a GAP function that can change its argument in
place, even if it is given an immutable argument. In particular,
`ImmutableMatrix` can change the representation of the matrix that it is
given.

For us, this is dangerous since matrices in different representations
are considered to be *not* equal in GAP, and they cannot in general be
multiplied. Therefore, if a matrix that belongs to a semigroup has its
representation changed in-place, then multiplication and equality
testing no longer necessarily work correctly in that semigroup. This is
what led to the issue in issue 788, where a semigroup of size 16 had 17
elements, because it had two copies of its identity in different
representations.