Update cartan.xml
This commit is for several changes related to the review of the code. In order of line appearance ion the code the changes relate to:
Changing a code reference to an attribute reference,
Changing a code reference to a more accurate attribute reference,
Correcting the syntax of an attribute reference,
An incomplete attempt to explain the order of the conjugacy classes as they would appear in a table,
A more accurate return suggested by the review,
A rewording to make it clearer that the field argument is optional,
A rewording and added space to separate the case,
Removed a comment refereeing to a future case implementation,
Added a new line for spacing.